fix: updated files directories
