+++ taskManagerController/src/main/kotlin/com/example/taskManagerController/TaskManagerControllerApplication.kt
@@ -1,4 +1,4 @@
-package com.example.taskManagerController
+package com.example.taskManagerController.service;
 
 import org.springframework.boot.SpringApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
