+++ taskManagerController/src/main/kotlin/com/example/taskManagerController/model/Task.kt
@@ -1,4 +1,4 @@
-package model;
+package com.example.taskManagerController.model;
 
 import java.time.LocalDate
 import java.util.UUID
