+++ taskManagerController/src/main/kotlin/com/example/taskManagerController/repository/TaskRepository.kt
@@ -1,6 +1,6 @@
-package repository;
+package com.example.taskManagerController.repository;
 
-import entity.TaskEntity;
+import com.example.taskManagerController.entity.TaskEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
